wsj mix scripts tidy up
