+++ egs/wsj_mix/asr1/cmd.sh
@@ -12,9 +12,9 @@
 
 export train_cmd="run.pl --mem 2G"
 export cuda_cmd="run.pl --mem 2G --gpu 1"
-#export decode_cmd="run.pl --mem 4G"
+export decode_cmd="run.pl --mem 4G"
 
 # JHU setup
 #export train_cmd="queue.pl --mem 2G"
 #export cuda_cmd="queue.pl --mem 2G --gpu 1 --config conf/gpu.conf"
-export decode_cmd="queue.pl --mem 4G"
+#export decode_cmd="queue.pl --mem 4G"
