+++ egs/wsj_mix/asr1/local/compute_perm_free_error.sh
