+++ egs/wsj_mix/asr1/local/json2trn.py
