+++ egs/wsj_mix/asr1/local/minpermwer.py
