+++ egs/wsj_mix/asr1/local/result2json.py
