+++ egs/wsj_mix/asr1/local/score_sclite.sh
