--- egs/wsj_mix/asr1/local/src/__init__.py
@@ -1,4 +1,0 @@
-#!/usr/bin/env python
-# encoding: utf-8
-
-
